feat!: remove dotenv support from Prisma Client
Since there's no `env()` in PSL anymore, we don't need to load
environment variables in Prisma Client.

But we also wanted to remove this logic regardless and have people load
their environment variables themselves, as it shouldn't be Prisma's
responsibility and Prisma shouldn't be mutating `process.env`.

This is already the case for the new generator, now we remove this
completely.

Ref: https://linear.app/prisma-company/issue/TML-1459/remove-processenv-loading-from-prismaclient
Ref: https://linear.app/prisma-company/issue/TML-1340/remove-env-from-psl
Closes: https://linear.app/prisma-company/issue/TML-309/fix-leaky-env-logic-→-stop-mutating-processenv